employee_attrition_project/pipeline_attrition/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread
pandas 
numpy
sqlalchemy
psycopg2-binary
python-dotenv
requests